Added a null check to CPlayerCamera (4132: Server crash after someone quits) Fixed up pointer arithmentic in CObjectSA::Get/SetHealth

+++ MTA10/game_sa/CObjectSA.cpp
@@ -237,12 +237,12 @@ void CObjectSA::SetScale( float faScale )
 
 void CObjectSA::SetHealth ( float fHealth )
 {
-    *(float *)(this->GetInterface() + 340) = fHealth;
+    *(float *)( (DWORD)this->GetInterface () + 340 ) = fHealth;
 }
 
 float CObjectSA::GetHealth ( void )
 {
-    return *(float *)(this->GetInterface() + 340);
+    return *(float *)( (DWORD)this->GetInterface () + 340 );
 }
 
 void CObjectSA::SetModelIndex ( unsigned long ulModel )
